# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/test/1": {
            "get": {
                "description": "测试 gin-swagger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试 gin-swagger"
                ],
                "summary": "测试 gin-swagger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Resp"
                        }
                    }
                }
            }
        },
        "/test/2/{id}": {
            "get": {
                "description": "测试 gin-swagger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试 gin-swagger"
                ],
                "summary": "测试 gin-swagger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "成功"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        }
    }
}